Tests: Refactoring
