+++ tests/entities/test_dataset.py
@@ -1,5 +1,5 @@
 import pytest
-from mlflow.entities.evaluation_dataset import EvaluationDataset
+from mlflow.entities import Dataset
 from mlflow.exceptions import RestException
 
 pytestmark = pytest.mark.slow
@@ -18,8 +18,9 @@ def test_client_dataset_basic(mlflow_server, mlflow_client, reset_database):
     """
     Validate basic MLflow dataset creation.
     """
+    dataset: Dataset
 
-    dataset: EvaluationDataset = mlflow_client.create_dataset("foo")
+    dataset = mlflow_client.create_dataset("foo")
     assert dataset.name == "foo"
 
     dataset = mlflow_client.get_dataset(dataset_id=dataset.dataset_id)
