+++ tests/infra/test_cli.py
