+++ tests/store/model_registry/test_api.py
