+++ tests/store/tracking/test_adapter.py
